clean code for cutree

+++ Source/Lib/Encoder/Codec/EbPictureControlSet.c
@@ -1340,7 +1340,7 @@ EbErrorType picture_parent_control_set_ctor(PictureParentControlSet *object_ptr,
     object_ptr->r0 = 0;
     object_ptr->is_720p_or_larger = AOMMIN(init_data_ptr->picture_width, init_data_ptr->picture_height) >= 720;
     EB_MALLOC_2D(object_ptr->ois_mb_results, picture_width_in_mb * picture_height_in_mb, 1);
-    EB_MALLOC_2D(object_ptr->cutree_stats, (picture_width_in_mb << (1 - object_ptr->is_720p_or_larger)) * (picture_height_in_mb << (1 - object_ptr->is_720p_or_larger)), 1);
+    EB_MALLOC_2D(object_ptr->cutree_stats, (uint32_t)((picture_width_in_mb << (1 - object_ptr->is_720p_or_larger)) * (picture_height_in_mb << (1 - object_ptr->is_720p_or_larger))), 1);
     EB_MALLOC_ARRAY(object_ptr->cutree_beta, object_ptr->sb_total_count);
 #if LAMBDA_SCALING
     EB_MALLOC_ARRAY(object_ptr->tpl_rdmult_scaling_factors, picture_width_in_mb * picture_height_in_mb);
